Changed color for *page info* in `IssueList` component (#3)
Previous behaviour: The page info has a wrong (lighter) color than in the design. The contrast is really low this way which is bad for accessibility. See the designs for the correct color.
Expected Behaviour: The text color of page info at the bottom of the issue list should match the design

Fixed 👍