On mobile the header always show the large logo. (#5)
Fixed this issue based on `isSidebarCollapsed` prop and `viewportWidth`.

Previous behaviour: When a user switches device orientation in a way that results in a change from desktop to mobile mode the small logo is shown instead of the large one.
Expected bahaviour: On mobile devices, the header should always show the large logo.

Fixed 👍